Move VMs to own namespace.
