+++ src/MrCapitalQ.FollowAlong/MainWindow.xaml
@@ -7,6 +7,7 @@
                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                   xmlns:controls="using:MrCapitalQ.FollowAlong.Controls"
                   xmlns:converters="using:MrCapitalQ.FollowAlong.Converters"
+                  xmlns:vm="using:MrCapitalQ.FollowAlong.ViewModels"
                   xmlns:winuiex="using:WinUIEx"
                   mc:Ignorable="d"
                   Width="800"
@@ -44,7 +45,7 @@
                         </ItemsPanelTemplate>
                     </GridView.ItemsPanel>
                     <GridView.ItemTemplate>
-                        <DataTemplate x:DataType="local:DisplayViewModel">
+                        <DataTemplate x:DataType="vm:DisplayViewModel">
                             <Border BorderBrush="Black"
                                     BorderThickness="4"
                                     CornerRadius="4"
